Renamed dao to datasource and FTPClientAssemblyDaoImplement to NCBIAssemblyDataSource
