+++ src/test/java/com/ebivariation/contigalias/datasource/AssemblyDataSourceTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.ebivariation.contigalias.dao;
+package com.ebivariation.contigalias.datasource;
 
 import com.ebivariation.contigalias.entities.AssemblyEntity;
 import com.ebivariation.contigalias.entities.ChromosomeEntity;
@@ -33,19 +33,19 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @SpringBootTest
-public class AssemblyDaoTest {
+public class AssemblyDataSourceTest {
 
     private static final String GCA_ACCESSION_HAVING_CHROMOSOMES = "GCA_000003055.3";
 
     private static final String GCF_ACCESSION_NO_CHROMOSOMES = "GCF_006125015.1";
 
     @Autowired
-    @Qualifier("ftpDao")
-    private AssemblyDao dao;
+    @Qualifier("NCBIDataSource")
+    private AssemblyDataSource dataSource;
 
     @Test
     public void getAssemblyByAccessionGCAHavingChromosomes() throws IOException{
-        Optional<AssemblyEntity> accession = dao.getAssemblyByAccession(GCA_ACCESSION_HAVING_CHROMOSOMES);
+        Optional<AssemblyEntity> accession = dataSource.getAssemblyByAccession(GCA_ACCESSION_HAVING_CHROMOSOMES);
         assertTrue(accession.isPresent());
         List<ChromosomeEntity> chromosomes = accession.get().getChromosomes();
         assertNotNull(chromosomes);
@@ -54,7 +54,7 @@ public class AssemblyDaoTest {
 
     @Test
     public void getAssemblyByAccessionGCFNoChromosomes() throws IOException {
-        Optional<AssemblyEntity> accession = dao.getAssemblyByAccession(GCF_ACCESSION_NO_CHROMOSOMES);
+        Optional<AssemblyEntity> accession = dataSource.getAssemblyByAccession(GCF_ACCESSION_NO_CHROMOSOMES);
         assertTrue(accession.isPresent());
         List<ChromosomeEntity> chromosomes = accession.get().getChromosomes();
         assertNull(chromosomes);
